Add direct download
Also puts it into your soundpad for u :D

+++ downloader/main.py
@@ -1,6 +1,9 @@
 from flask import Blueprint, render_template, request, redirect, url_for
 from yt_dlp import YoutubeDL
 from youtube_search import YoutubeSearch
+from urllib.parse import quote_plus
+from .soundpad import soundpad
+import time
 import os
 
 main = Blueprint("main", __name__)
@@ -37,15 +40,20 @@ def search_get():
         return render_template("search.html", title="Search", videos=videos)
 
 
-@main.route("/download/<video_id>/<download_slot>", methods=["GET", "POST"])
-def download(video_id, download_slot):
-    try:
-        os.remove(f"/audio/{download_slot}.mp3")
-    except FileNotFoundError:
-        pass
+@main.route("/download/<video_id>/<download_slot>/<title>", methods=["GET", "POST"])
+def download(video_id, download_slot, title):
+    if download_slot.isnumeric():
+        try:
+            os.remove(f"/audio/{download_slot}.mp3")
+        except FileNotFoundError:
+            pass
+        options = ydl_options
+        options["outtmpl"] = f"/audio/{download_slot}.%(ext)s"
+    else:
+        title = quote_plus(title)
+        options = ydl_options
+        options["outtmpl"] = f"/audio/{title}.%(ext)s"
 
-    options = ydl_options
-    options["outtmpl"] = f"/audio/{download_slot}.%(ext)s"
     success = True
     try:
         with YoutubeDL(options) as ydl:
@@ -53,6 +61,10 @@ def download(video_id, download_slot):
     except Exception as e:
         success = False
         print(e)
+
+    if not download_slot.isnumeric():
+        time.sleep(1)
+        soundpad.addSound(f"{os.getcwd()}/audio/{title}.mp3")
 
     search = request.args.get("search")
     if search is None:
